binary_sensor.py --> __init__.py
